LayoutTests/platform/mac/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-video-element/video_timeupdate_on_seek-expected.txt


PASS timeupdate is emitted after a seek before the data is received: mp4.
